Clean up main
